requirements-dev.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements-dev.in
#
asgiref==3.7.2
    # via django
build==1.0.3
    # via pip-tools
certifi==2023.7.22
    # via requests
cfgv==3.4.0
    # via pre-commit
charset-normalizer==3.2.0
    # via requests
click==8.1.7
    # via pip-tools
distlib==0.3.7
    # via virtualenv
django==4.2.5
    # via
    #   -r requirements.in
    #   django-debug-toolbar
    #   django-filter
    #   django-htmx
django-debug-toolbar==4.2.0
    # via -r requirements-dev.in
django-filter==23.2
    # via -r requirements.in
django-htmx==1.16.0
    # via -r requirements.in
filelock==3.12.4
    # via virtualenv
gunicorn==21.2.0
    # via -r requirements.in
identify==2.5.30
    # via pre-commit
idna==3.4
    # via requests
nodeenv==1.8.0
    # via pre-commit
packaging==23.1
    # via
    #   build
    #   gunicorn
pillow==10.0.1
    # via -r requirements.in
pip-tools==7.3.0
    # via -r requirements-dev.in
platformdirs==3.11.0
    # via virtualenv
pre-commit==3.5.0
    # via -r requirements-dev.in
psycopg2-binary==2.9.7
    # via -r requirements.in
pyproject-hooks==1.0.0
    # via build
python-dotenv==1.0.0
    # via -r requirements.in
pyyaml==6.0.1
    # via pre-commit
requests==2.31.0
    # via -r requirements.in
sqlparse==0.4.4
    # via
    #   django
    #   django-debug-toolbar
tomli==2.0.1
    # via
    #   build
    #   pip-tools
    #   pyproject-hooks
typing-extensions==4.8.0
    # via asgiref
urllib3==2.0.4
    # via requests
virtualenv==20.24.6
    # via pre-commit
wheel==0.41.2
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
